+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Stop recording `unload` event
 - Stop recording the whole `attributes` object from html elements
 
 ### Fixed